third_party/blink/web_tests/platform/generic/http/tests/xmlhttprequest/getResponseHeader-expected.txt

CONSOLE ERROR: Refused to get unsafe header "SeT-COoKie"
CONSOLE ERROR: Refused to get unsafe header "sEt-coOkIE2"
CONSOLE ERROR: Refused to get unsafe header "SeT-COoKie"
CONSOLE ERROR: Refused to get unsafe header "sEt-coOkIE2"
CONSOLE ERROR: Refused to get unsafe header "SeT-COoKie"
CONSOLE ERROR: Refused to get unsafe header "sEt-coOkIE2"
Test the required behavior of XMLHttpRequest.getResponseHeader()

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS {state: 0}; headerValue = xhr.getResponseHeader('Content-Type'); did not throw exception.
PASS headerValue is null
PASS {state: 1}; headerValue = xhr.getResponseHeader('Content-Type'); did not throw exception.
PASS headerValue is null
PASS Header 'Content-Type' is null in ready state 1.
PASS xhr.open("GET", "resources/1251.html", true); did not throw exception.
PASS {state: 1}; headerValue = xhr.getResponseHeader('Content-Type'); did not throw exception.
PASS headerValue is null
PASS xhr.send(null); did not throw exception.
PASS {state: 2}; headerValue = xhr.getResponseHeader('Content-Type'); did not throw exception.
PASS headerValue is non-null.
PASS {state: 2}; headerValue = xhr.getResponseHeader('SeT-COoKie'); did not throw exception.
PASS headerValue is null
PASS {state: 2}; headerValue = xhr.getResponseHeader('sEt-coOkIE2'); did not throw exception.
PASS headerValue is null
PASS {state: 2}; headerValue = xhr.getResponseHeader('xxx-mytest-headerabc'); did not throw exception.
PASS headerValue is null
PASS {state: 2}; headerValue = xhr.getResponseHeader('null'); did not throw exception.
PASS headerValue is null
PASS {state: 2}; headerValue = xhr.getResponseHeader('Content-Type:'); did not throw exception.
PASS headerValue is null
PASS {state: 3}; headerValue = xhr.getResponseHeader('Content-Type'); did not throw exception.
PASS headerValue is non-null.
PASS result is savedHeader
PASS {state: 3}; headerValue = xhr.getResponseHeader('SeT-COoKie'); did not throw exception.
PASS headerValue is null
PASS {state: 3}; headerValue = xhr.getResponseHeader('sEt-coOkIE2'); did not throw exception.
PASS headerValue is null
PASS {state: 3}; headerValue = xhr.getResponseHeader('xxx-mytest-headerabc'); did not throw exception.
PASS headerValue is null
PASS {state: 3}; headerValue = xhr.getResponseHeader('null'); did not throw exception.
PASS headerValue is null
PASS {state: 3}; headerValue = xhr.getResponseHeader('Content-Type:'); did not throw exception.
PASS headerValue is null
PASS {state: 4}; headerValue = xhr.getResponseHeader('Content-Type'); did not throw exception.
PASS headerValue is non-null.
PASS result is savedHeader
PASS {state: 4}; headerValue = xhr.getResponseHeader('SeT-COoKie'); did not throw exception.
PASS headerValue is null
PASS {state: 4}; headerValue = xhr.getResponseHeader('sEt-coOkIE2'); did not throw exception.
PASS headerValue is null
PASS {state: 4}; headerValue = xhr.getResponseHeader('xxx-mytest-headerabc'); did not throw exception.
PASS headerValue is null
PASS {state: 4}; headerValue = xhr.getResponseHeader('null'); did not throw exception.
PASS headerValue is null
PASS {state: 4}; headerValue = xhr.getResponseHeader('Content-Type:'); did not throw exception.
PASS headerValue is null
PASS successfullyParsed is true

TEST COMPLETE
